- Updated the expired error text - Refactored the Deferred Installation component
